% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning.R
\name{datathin}
\alias{datathin}
\title{Takes a dataset (scalar, vector, or matrix) and returns a training set and a test set that sum to the original data matrix.}
\usage{
datathin(data, family, K = 2, epsilon = NULL, arg = NULL)
}
\description{
Takes a dataset (scalar, vector, or matrix) and returns a training set and a test set that sum to the original data matrix.
}
