% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datathin.R
\name{multithin}
\alias{multithin}
\title{Takes a dataset (scalar, vector, or matrix???) and returns multiple folds of data that sum to the original data matrix.}
\usage{
multithin(data, family, nfolds = 5, arg = NULL)
}
\arguments{
\item{family}{The name of the distribution of the data. Options are "poisson", "negative binomial", "normal" (equivalently "gaussian),
"binomial", "exponential", or "gamma".}

\item{nfolds}{The number of folds to create from the data.}

\item{arg}{The extra parameter that must be known in order to split. Not needed for Poisson, but needed for all other distributions.
Should be a scalar or should match dimensions of X.}

\item{X}{A scalar, vector, or matrix of data.}
}
\description{
Takes a dataset (scalar, vector, or matrix???) and returns multiple folds of data that sum to the original data matrix.
}
